:lipstick: Move balance to portfolio and show all balance data there
