+++ app/[usernameSlug]/page.tsx
@@ -33,23 +33,10 @@ export default async function UserProfilePage(props: {
   const isOwnProfile = user?.id === profile?.id
   const txns = await getTxnAndProjectsByUser(supabase, profile.id)
   const investments = await compileInvestments(txns, profile.id)
-  const balance = calculateUserBalance(txns, profile.id)
-  const withdrawBalance = calculateWithdrawBalance(
-    txns,
-    bids,
-    profile.id,
-    balance,
-    profile.accreditation_status
-  )
 
   return (
     <div className="flex flex-col p-5">
-      <ProfileHeader
-        profile={profile}
-        isOwnProfile={isOwnProfile}
-        balance={balance}
-        withdrawBalance={withdrawBalance}
-      />
+      <ProfileHeader profile={profile} isOwnProfile={isOwnProfile} />
       <div className="flex flex-col gap-10">
         <ProfileTabs
           isOwnProfile={isOwnProfile}
@@ -57,6 +44,7 @@ export default async function UserProfilePage(props: {
           projects={projects}
           bids={bids}
           investments={investments}
+          txns={txns}
         />
         {isOwnProfile && (
           <div className="mt-5 flex justify-center">
@@ -100,116 +88,4 @@ async function compileInvestments(txns: TxnAndProject[], userId: string) {
     }
   })
   return investments as Investment[]
-}
-
-function calculateWithdrawBalance(
-  txns: Txn[],
-  bids: Bid[],
-  userId: string,
-  balance: number,
-  accreditationStatus: boolean
-) {
-  let nonWithdrawBalance = 0
-  const sortedTxns = sortBy(txns, 'created_at')
-  sortedTxns.forEach((txn) => {
-    const txnType = categorizeTxn(txn, userId)
-    if (txnType === 'non-dollar') return
-    if (txn.to_id === userId) {
-      if (!txnWithdrawable(txnType, accreditationStatus)) {
-        nonWithdrawBalance += txn.amount
-      }
-    } else {
-      if (txnCharitable(txnType, accreditationStatus)) {
-        nonWithdrawBalance -= txn.amount
-      }
-    }
-    nonWithdrawBalance = Math.max(nonWithdrawBalance, 0)
-  })
-  bids.forEach((bid) => {
-    if (bid.status === 'pending' && bid.type === 'buy') {
-      nonWithdrawBalance += bid.amount
-    }
-  })
-
-  return Math.max(balance - nonWithdrawBalance, 0)
-}
-
-type TxnType =
-  | 'incoming cash transfer'
-  | 'outgoing cash transfer'
-  | 'share purchase'
-  | 'share sale'
-  | 'own-project share sale'
-  | 'withdraw'
-  | 'deposit'
-  | 'incoming project donation'
-  | 'outgoing project donation'
-  | 'non-dollar'
-
-function categorizeTxn(txn: TxnAndProject, userId: string) {
-  if (txn.token === 'USD') {
-    if (txn.to_id === userId) {
-      if (txn.project) {
-        if (txn.projects?.creator === userId) {
-          if (txn.bundle) {
-            return 'own-project share sale'
-          } else {
-            return 'incoming project donation'
-          }
-        } else {
-          return 'share sale'
-        }
-      } else {
-        if (txn.from_id === BANK_ID) {
-          return 'deposit'
-        } else {
-          return 'incoming cash transfer'
-        }
-      }
-    } else {
-      if (txn.project) {
-        if (txn.bundle) {
-          return 'share purchase'
-        } else {
-          return 'outgoing project donation'
-        }
-      } else {
-        if (txn.to_id === BANK_ID) {
-          return 'withdraw'
-        } else {
-          return 'outgoing cash transfer'
-        }
-      }
-    }
-  } else {
-    return 'non-dollar'
-  }
-}
-
-// For incoming txns only
-function txnWithdrawable(txnType: TxnType, accreditationStatus: boolean) {
-  if (txnType === 'incoming cash transfer') {
-    return false
-  }
-  if (
-    txnType === 'share sale' ||
-    (txnType === 'deposit' && !accreditationStatus)
-  ) {
-    return false
-  }
-  return true
-}
-
-// For outgoing txns only
-function txnCharitable(txnType: TxnType, accreditationStatus: boolean) {
-  if (
-    txnType === 'outgoing cash transfer' ||
-    txnType === 'outgoing project donation'
-  ) {
-    return true
-  }
-  if (txnType === 'share purchase' && !accreditationStatus) {
-    return true
-  }
-  return false
 }
