+++ app/[usernameSlug]/profile-header.tsx
@@ -8,17 +8,18 @@ import { InvestorTypeTag } from '@/components/tags'
 import { addHttpToUrl } from '@/utils/formatting'
 import { Row } from '@/components/layout/row'
 import { Col } from '@/components/layout/col'
+import { useRef } from 'react'
 
 type Profile = Database['public']['Tables']['profiles']['Row']
 
 export function ProfileHeader(props: {
   profile: Profile
   isOwnProfile: boolean
-  balance: number
-  withdrawBalance: number
 }) {
-  const { profile, isOwnProfile, balance, withdrawBalance } = props
+  const { profile, isOwnProfile } = props
   const website = addHttpToUrl(profile.website ?? '')
+  // Doesn't change when you change window size
+  const windowWidth = useRef([window.innerWidth, window.innerHeight]).current[0]
   return (
     <div className="flex flex-col gap-3">
       <div className="flex">
@@ -26,19 +27,24 @@ export function ProfileHeader(props: {
           username={profile.username}
           avatarUrl={profile.avatar_url}
           noLink
-          size={24}
+          size={windowWidth > 640 ? 24 : undefined}
         />
         {isOwnProfile && (
-          <div className="relative top-14 right-6 h-10 w-10 rounded-full bg-orange-400 hover:bg-orange-500">
+          <div className="relative top-8 right-4 h-5 w-5 rounded-full bg-orange-400 hover:bg-orange-500 sm:top-14 sm:right-6 sm:h-10 sm:w-10">
             <Link href="/edit-profile">
-              <PencilIcon className="h-10 w-10 p-2" aria-hidden />
+              <PencilIcon
+                className="h-5 w-5 p-1 sm:h-10 sm:w-10 sm:p-2"
+                aria-hidden
+              />
             </Link>
           </div>
         )}
         <Col className="w-full">
           <Row className="justify-between">
-            <Col className="ml-3">
-              <div className="text-3xl font-bold">{profile.full_name}</div>
+            <Col>
+              <div className="text-lg font-bold leading-tight sm:text-2xl">
+                {profile.full_name}
+              </div>
               <Row className="mt-1 flex-wrap gap-2 text-gray-500">
                 <p>@{profile.username}</p>
                 <InvestorTypeTag
@@ -48,13 +54,6 @@ export function ProfileHeader(props: {
                 />
               </Row>
             </Col>
-            {isOwnProfile && (
-              <BalanceBox
-                balance={balance}
-                withdrawBalance={withdrawBalance}
-                accredited={profile.accreditation_status}
-              />
-            )}
           </Row>
         </Col>
       </div>
